# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AudienceMember": {
        "properties": {
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "consent_state": {
            "title": "Consent State",
            "type": "string"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "fit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fit"
          }
        },
        "required": [
          "candidate_id",
          "external_ref",
          "consent_state",
          "fit"
        ],
        "title": "AudienceMember",
        "type": "object"
      },
      "AudienceOut": {
        "description": "Who is eligible for outreach: opted-in only (CAN-SPAM); excluded counts logged.",
        "properties": {
          "eligible": {
            "title": "Eligible",
            "type": "integer"
          },
          "excluded_pending": {
            "title": "Excluded Pending",
            "type": "integer"
          },
          "excluded_unsubscribed": {
            "title": "Excluded Unsubscribed",
            "type": "integer"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/AudienceMember"
            },
            "title": "Members",
            "type": "array"
          },
          "requisition_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requisition Id"
          }
        },
        "required": [
          "requisition_id",
          "eligible",
          "excluded_unsubscribed",
          "excluded_pending",
          "members"
        ],
        "title": "AudienceOut",
        "type": "object"
      },
...
      "DashboardStats": {
        "properties": {
          "candidates": {
            "title": "Candidates",
            "type": "integer"
          },
          "open_requisitions": {
            "title": "Open Requisitions",
            "type": "integer"
          },
          "opted_in": {
            "title": "Opted In",
            "type": "integer"
          },
          "proposals": {
            "title": "Proposals",
            "type": "integer"
          },
          "requisitions": {
            "title": "Requisitions",
            "type": "integer"
          }
        },
        "required": [
          "candidates",
          "requisitions",
          "open_requisitions",
          "proposals",
          "opted_in"
        ],
        "title": "DashboardStats",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OutreachStatsOut": {
        "properties": {
          "opted_in": {
            "title": "Opted In",
            "type": "integer"
          },
          "pending": {
            "title": "Pending",
            "type": "integer"
          },
          "total_candidates": {
            "title": "Total Candidates",
            "type": "integer"
          },
          "unsubscribed": {
            "title": "Unsubscribed",
            "type": "integer"
          }
        },
        "required": [
          "total_candidates",
          "opted_in",
          "unsubscribed",
          "pending"
        ],
        "title": "OutreachStatsOut",
        "type": "object"
      },
      "PlanOut": {
        "description": "Plan placeholder. Pricing + billing are owner/counsel-gated (\u2696\ufe0f) and not yet wired.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "price": {
            "title": "Price",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "name",
          "price",
          "unit",
          "note"
        ],
        "title": "PlanOut",
        "type": "object"
      },
      "ProposalCreate": {
        "properties": {
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "outcome": {
            "default": "proposed",
            "title": "Outcome",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "requisition_id": {
            "title": "Requisition Id",
            "type": "string"
          }
        },
        "required": [
          "candidate_id",
          "requisition_id"
        ],
        "title": "ProposalCreate",
        "type": "object"
      },
      "ProposalOut": {
        "properties": {
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "requisition_id": {
            "title": "Requisition Id",
            "type": "string"
          },
          "requisition_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requisition Title"
          }
        },
        "required": [
          "id",
          "candidate_id",
          "requisition_id",
          "requisition_title",
          "outcome",
          "reason",
          "decided_at",
          "created_at"
        ],
        "title": "ProposalOut",
        "type": "object"
      },
...
      "TeamMemberOut": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "status"
        ],
        "title": "TeamMemberOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Internal AI/services API. Phase 1 product surface over the Phase 0 spine.",
    "title": "ManFriday Internal API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/dashboard/stats": {
      "get": {
        "description": "Headline counts for the org's dashboard.",
        "operationId": "getDashboardStats",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStats"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard Stats",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/outreach/audience": {
      "get": {
        "description": "Who is eligible for outreach: opted-in candidates, optionally fitting a requisition.\n\nWithout ``requisition_id`` it's every opted-in candidate. With one, only those whose\ntransparent skill-overlap fit \u2265 ``min_fit``. Excluded (unsubscribed / pending) are\ncounted, never silently dropped (CAN-SPAM consent).",
        "operationId": "getAudience",
        "parameters": [
          {
            "in": "query",
            "name": "requisition_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requisition Id"
            }
          },
          {
            "in": "query",
            "name": "min_fit",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 100,
              "minimum": 0,
              "title": "Min Fit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudienceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Audience",
        "tags": [
          "outreach"
        ]
      }
    },
    "/outreach/stats": {
      "get": {
        "description": "Consent breakdown across the org's candidates.",
        "operationId": "getOutreachStats",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutreachStatsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Outreach Stats",
        "tags": [
          "outreach"
        ]
      }
    },
    "/plan": {
      "get": {
        "description": "Plan placeholder (pricing/billing are owner/counsel-gated and not yet wired).",
        "operationId": "getPlan",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Plan",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/proposals": {
      "get": {
        "description": "List proposals, optionally filtered to one candidate (their within-org history).",
        "operationId": "listProposals",
        "parameters": [
          {
            "in": "query",
            "name": "candidate_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProposalOut"
                  },
                  "title": "Response Listproposals",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Proposals",
        "tags": [
          "proposals"
        ]
      },
      "post": {
        "description": "Record that a candidate was proposed to a requisition (human action, audited).",
        "operationId": "createProposal",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Proposal",
        "tags": [
          "proposals"
        ]
      }
    },
...
    "/team": {
      "get": {
        "description": "List the org's users (the Settings \u2192 Team view).",
        "operationId": "listTeamMembers",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TeamMemberOut"
                  },
                  "title": "Response Listteammembers",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Team",
        "tags": [
          "dashboard"
        ]
      }
    },
...
  }
}